Ben/requirements.txt
numpy==1.26.1
opencv-python==4.8.1.78
pygame==2.5.2
pyserial==3.5
